refactor code project ts
