+++ BabylonJS/src/BasicScreen.ts
@@ -1,0 +1,66 @@
+import { GridMaterial } from "@babylonjs/materials/grid";
+import * as BABYLON from '@babylonjs/core'
+
+import {scene} from './Enviroment'
+
+export function InitGround(){
+    var groundMaterial = new GridMaterial("groundMaterial", scene);
+    groundMaterial.gridRatio = 1;
+    groundMaterial.backFaceCulling = false;
+    groundMaterial.mainColor = new BABYLON.Color3(1, 1, 1);
+    groundMaterial.lineColor = new BABYLON.Color3(1.0, 1.0, 1.0);
+    groundMaterial.opacity = 0.5;
+
+    var groundz = BABYLON.MeshBuilder.CreateGround("groundz", { width: 100, height: 100 }, scene);
+    groundz.material = groundMaterial;
+    // groundz.material.mainColor = new BABYLON.Color3(0,0,1);
+
+    var groundx = BABYLON.MeshBuilder.CreateGround("groundx", { width: 100, height: 100 }, scene);
+    groundx.rotation.x = Math.PI / 2;
+    groundx.material = groundMaterial;
+    // groundz.material.mainColor = new BABYLON.Color3(1,0,0);
+
+    var groundy = BABYLON.MeshBuilder.CreateGround("groundy", { width: 100, height: 100 }, scene);
+    groundy.rotation.z = Math.PI / 2;
+    groundy.material = groundMaterial;
+
+}
+
+///////////////// Show Axis ////////////////////////
+export function showAxis(size) {
+    var makeTextPlane = function(text, color, size) {
+        var dynamicTexture = new BABYLON.DynamicTexture("DynamicTexture", 50, scene, true);
+        dynamicTexture.hasAlpha = true;
+        dynamicTexture.drawText(text, 5, 40, "bold 36px Arial", color, "transparent", true);
+        var plane = BABYLON.Mesh.CreatePlane("TextPlane", size, scene, true);
+        plane.material = new BABYLON.StandardMaterial("TextPlaneMaterial", scene);
+        plane.material.backFaceCulling = false;
+        (plane.material as BABYLON.StandardMaterial).specularColor = new BABYLON.Color3(0, 0, 0);
+        (plane.material as BABYLON.StandardMaterial).diffuseTexture = dynamicTexture;
+        return plane;
+    };
+
+    var axisX = BABYLON.Mesh.CreateLines("axisX", [
+        BABYLON.Vector3.Zero(), new BABYLON.Vector3(size, 0, 0), new BABYLON.Vector3(size * 0.95, 0.05 * size, 0),
+        new BABYLON.Vector3(size, 0, 0), new BABYLON.Vector3(size * 0.95, -0.05 * size, 0)
+    ], scene);
+    axisX.color = new BABYLON.Color3(1, 0, 0);
+    var xChar = makeTextPlane("X", "red", size / 10);
+    xChar.position = new BABYLON.Vector3(size, -0.1 * size, 0);
+    var axisY = BABYLON.Mesh.CreateLines("axisY", [
+        BABYLON.Vector3.Zero(), new BABYLON.Vector3(0, size, 0), new BABYLON.Vector3(-0.05 * size, size * 0.95, 0),
+        new BABYLON.Vector3(0, size, 0), new BABYLON.Vector3(0.05 * size, size * 0.95, 0)
+    ], scene);
+    axisY.color = new BABYLON.Color3(0, 1, 0);
+    var yChar = makeTextPlane("Y", "green", size / 10);
+    yChar.position = new BABYLON.Vector3(0, 0.9 * size, 0);
+    var axisZ = BABYLON.Mesh.CreateLines("axisZ", [
+        BABYLON.Vector3.Zero(), new BABYLON.Vector3(0, 0, size), new BABYLON.Vector3(0, -0.05 * size, size * 0.95),
+        new BABYLON.Vector3(0, 0, size), new BABYLON.Vector3(0, 0.05 * size, size * 0.95)
+    ], scene);
+    axisZ.color = new BABYLON.Color3(0, 0, 1);
+    var zChar = makeTextPlane("Z", "blue", size / 10);
+    zChar.position = new BABYLON.Vector3(0, 0.05 * size, 0.9 * size);
+};
+
+// showAxis(10);
