+++ BabylonJS/src/Point.ts
@@ -1,15 +1,26 @@
 import * as BABYLON from "@babylonjs/core"
 
-function Point(){
-    var name;
-    var position;
-    var mesh;
+import {gizmoManager, scene, hl} from './Enviroment'
+ 
+export var pointMaster = BABYLON.MeshBuilder.CreateSphere("point", { diameter: 0.2 }, scene);
+    gizmoManager.attachableMeshes.push(pointMaster);
+    pointMaster.position.y = 1000;
+    var pointMat = new BABYLON.StandardMaterial("pointMat", scene);
+    pointMat.diffuseColor = BABYLON.Color3.Red();
+    pointMaster.material = pointMat;
+
+export class Point{
+    name: string;
+    position: BABYLON.Vector3;
+    mesh: BABYLON.Mesh;
+    constructor(name, position, mesh){
+        this.name = name;
+        this.position = position;
+        this.mesh = mesh;
+    }
 }
-export function CreatePoint(position,point,gizmoManager,hl,scene,isSelectMode,checkInSelectedMeshes) {
-    const _point=new Point();
-    _point.name="point";
-    _point.position=position;
-    _point.mesh=point.clone("point");
+export function CreatePoint(position,isSelectMode,checkInSelectedMeshes) {
+    const _point= new Point("point", position, pointMaster.clone("point"));
     gizmoManager.attachableMeshes.push(_point.mesh);
     // p.isVisible = true;
     _point.mesh.position = _point.position;
